requirements.txt
anyio==3.7.1
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
contourpy==1.1.0
crispy-bootstrap5==0.7
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-crispy-forms==2.0
EasyProcess==1.1
entrypoint2==1.1
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
fonttools==4.41.1
fqdn==1.5.1
html5lib==1.1
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
MouseInfo==0.1.3
mss==9.0.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.1
notebook_shim==0.2.3
numpy==1.25.2
oscrypto==1.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.15.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyjokes==0.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pypdf==3.14.0
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
reportlab==3.6.13
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.5.0
svglib==1.5.1
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
uritools==4.0.1
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wikipedia==1.4.0
xhtml2pdf==0.2.11
